chore: 🔧 rename SvgComponent to SvgWrapper
