+++ apps/reaction-speed-test/components/measurement/SvgWrapper.tsx
